ZIP files now use DEFLATE

+++ FileCompression/CRC32.cpp
@@ -5,10 +5,10 @@
 CRC32::CRC32()
 {
     // Initialize lookup table
-    for(int i = 0; i < 256; i ++)
+    for (int i = 0; i < 256; i++)
     {
         unsigned crc = i;
-        for(int j = 0; j < 8; j ++)
+        for (int j = 0; j < 8; j++)
         {
             crc = (crc & 1) ? (crc >> 1) ^ reversed_polynomial_ : (crc >> 1);
         }
@@ -19,8 +19,8 @@ CRC32::CRC32()
 unsigned CRC32::compute(const char* bytes, int length) const
 {
     unsigned crc = 0xffffffff;
-    while(length--)
-        crc = (crc >> 8) ^ lookup_table_[(crc & 0xff) ^ *bytes++];
+    while (length--)
+        crc = (crc >> 8) ^ lookup_table_[(crc & 0xff) ^ (Byte) (*bytes++)];
 
     return ~crc;
 }
